+++ include/tudocomp/compressors/lz77Aprox/stackoverflow_hash.hpp
@@ -19,7 +19,7 @@ namespace tdc{
         ~stackoverflow_hash(){};
 
 
-        long long make_hash(len_t start, len_t size,io::InputView &input_view){
+        uint64_t make_hash(len_t start, len_t size,io::InputView &input_view){
 
 
             long long hash =0;
